% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_taxonomy.R
\name{bf_tax_looks_non_taxon}
\alias{bf_tax_looks_non_taxon}
\title{Detect likely non-taxonomic strings}
\usage{
bf_tax_looks_non_taxon(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
A logical vector with the same length as \code{x}. Values are \code{TRUE} for
missing values, placeholders, measurement-like strings, environmental
variables or other inputs that are unlikely to be biological taxon names.
Values are \code{FALSE} for strings that remain plausible taxonomic names after
the conservative screening rules.

Detect strings that are unlikely to be biological taxon names
}
\description{
Flags common environmental variables, habitat descriptors, measurement names,
and other strings that should not be submitted to taxonomy APIs.
}
