% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{package}
\name{biofetchR-package}
\alias{biofetchR}
\alias{biofetchR-package}
\title{Internal package imports and global-variable declarations}
\description{
These declarations prevent false-positive R CMD check notes for
non-standard evaluation columns used in dplyr pipelines and spatial helpers.
}
\author{
\strong{Maintainer}: Darren Stuart \email{dstuart04@qub.ac.uk}

}
\keyword{internal}
